Reconsidered allowing Number inserts - now allows double only

+++ src/main/java/com/bigml/histogram/GroupTarget.java
@@ -16,7 +16,7 @@ public class GroupTarget extends Target<GroupTarget> {
     for (Object value : values) {
       Target target;
       if (value instanceof Number) {
-        target = new NumericTarget((Number) value);
+        target = new NumericTarget(((Number) value).doubleValue());
       } else {
         target = new CategoricalTarget(value);
       }
